refactor ballot compact
